In this commit, I refactored the entire project.
