--- SY2024-CORE-A-master/.pathplanner/settings.json
@@ -1,12 +1,0 @@
-{
-  "robotWidth": 0.9,
-  "robotLength": 0.9,
-  "holonomicMode": true,
-  "pathFolders": [],
-  "autoFolders": [],
-  "defaultMaxVel": 3.0,
-  "defaultMaxAccel": 3.0,
-  "defaultMaxAngVel": 540.0,
-  "defaultMaxAngAccel": 720.0,
-  "maxModuleSpeed": 4.5
-}
