--- SY2024-CORE-A-master/.wpilib/wpilib_preferences.json
@@ -1,6 +1,0 @@
-{
-    "enableCppIntellisense": false,
-    "currentLanguage": "java",
-    "projectYear": "2024",
-    "teamNumber": 4645
-}
