--- SY2024-CORE-A-master/README.md
@@ -1,25 +1,0 @@
-# SY2024 Core A
-
-Core A code for the Chicago Style Bot Dogs during SY2024, for the 2024 Crescendo game.
-
-## Documentation
-JavaDocs are included in most source files.
-
-Electronics and Pneumatics Diagrams & Documentation: https://docs.google.com/document/d/1Hy8igRGvVcIlELJ_j0sovlJ61RMXrC3zumiDmch_16E/edit
-
-## Usage
-This is installed just like any other WPILib 2024 project.
-
-To deploy, press CTRL + SHIFT + P and then select "WPILib: Deploy Robot Code".
-
-Make sure that your roboRIO is up-to-date.
-
-## Contributing
-Feel free to open a PR or an issue, but talk with a maintainer first.
-
-All significant changes should be:
-- Discussed with a maintainer
-- Developed on a separate branch
-- Reviewed and, hopefully, merged
-
-Small changes that don't affect the robot's functionality, like fixing typos, should generally be added more quickly.
