--- SY2024-CORE-A-master/src/main/deploy/pathplanner/autos/Auto Move & Turn.auto
@@ -1,25 +1,0 @@
-{
-  "version": 1.0,
-  "startingPose": {
-    "position": {
-      "x": 2,
-      "y": 2
-    },
-    "rotation": 0
-  },
-  "command": {
-    "type": "sequential",
-    "data": {
-      "commands": [
-        {
-          "type": "path",
-          "data": {
-            "pathName": "Path Move & Turn"
-          }
-        }
-      ]
-    }
-  },
-  "folder": null,
-  "choreoAuto": false
-}
