--- SY2024-CORE-A-master/src/main/deploy/pathplanner/paths/Path Complex.path
@@ -1,65 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 4.228906583554568,
-        "y": 6.690138362483131
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 5.228906583554571,
-        "y": 6.190138362483131
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 1.5366277154892876,
-        "y": 2.7579224442941834
-      },
-      "prevControl": {
-        "x": 0.5366277154892876,
-        "y": 3.7579224442941825
-      },
-      "nextControl": {
-        "x": 2.5366277154892907,
-        "y": 1.7579224442941834
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 4.122627721695233,
-        "y": 1.146149587238093
-      },
-      "prevControl": {
-        "x": 4.1369545915404835,
-        "y": 1.1389861523203793
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [],
-  "globalConstraints": {
-    "maxVelocity": 3.0,
-    "maxAcceleration": 3.0,
-    "maxAngularVelocity": 540.0,
-    "maxAngularAcceleration": 720.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 0,
-    "rotateFast": false
-  },
-  "reversed": false,
-  "folder": null,
-  "previewStartingState": null,
-  "useDefaultConstraints": false
-}
