--- SY2024-CORE-A-master/src/main/deploy/pathplanner/paths/Path Move & Turn.path
@@ -1,65 +1,0 @@
-{
-  "version": 1.0,
-  "waypoints": [
-    {
-      "anchor": {
-        "x": 2.0,
-        "y": 2.0
-      },
-      "prevControl": null,
-      "nextControl": {
-        "x": 3.051698678063767,
-        "y": 2.0
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 2.82,
-        "y": 2.0
-      },
-      "prevControl": {
-        "x": 2.933785330502656,
-        "y": 2.0
-      },
-      "nextControl": {
-        "x": 2.7062146694973435,
-        "y": 2.0
-      },
-      "isLocked": false,
-      "linkedName": null
-    },
-    {
-      "anchor": {
-        "x": 4.0,
-        "y": 2.0
-      },
-      "prevControl": {
-        "x": 2.4793093674254436,
-        "y": 2.0
-      },
-      "nextControl": null,
-      "isLocked": false,
-      "linkedName": null
-    }
-  ],
-  "rotationTargets": [],
-  "constraintZones": [],
-  "eventMarkers": [],
-  "globalConstraints": {
-    "maxVelocity": 3.0,
-    "maxAcceleration": 3.0,
-    "maxAngularVelocity": 540.0,
-    "maxAngularAcceleration": 720.0
-  },
-  "goalEndState": {
-    "velocity": 0,
-    "rotation": 90.0,
-    "rotateFast": false
-  },
-  "reversed": false,
-  "folder": null,
-  "previewStartingState": null,
-  "useDefaultConstraints": false
-}
