--- SY2024-CORE-A-master/src/main/java/frc/lib/config/SwerveModuleConstants.java
@@ -1,26 +1,0 @@
-package frc.lib.config;
-
-import edu.wpi.first.math.geometry.Rotation2d;
-
-public class SwerveModuleConstants {
-  public final int driveMotorID;
-  public final int angleMotorID;
-  public final int cancoderID;
-  public final Rotation2d angleOffset;
-
-  /**
-   * Swerve Module Constants to be used when creating swerve modules.
-   *
-   * @param driveMotorID
-   * @param angleMotorID
-   * @param canCoderID
-   * @param angleOffset
-   */
-  public SwerveModuleConstants(
-      int driveMotorID, int angleMotorID, int canCoderID, Rotation2d angleOffset) {
-    this.driveMotorID = driveMotorID;
-    this.angleMotorID = angleMotorID;
-    this.cancoderID = canCoderID;
-    this.angleOffset = angleOffset;
-  }
-}
