--- SY2024-CORE-A-master/src/main/java/frc/lib/util/AprilTag.java
@@ -1,193 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.lib.util;
-
-import java.beans.Visibility;
-
-import org.opencv.core.Mat.Tuple2;
-
-import edu.wpi.first.math.Pair;
-import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.geometry.Rotation3d;
-import edu.wpi.first.math.geometry.Transform3d;
-import edu.wpi.first.math.geometry.Translation3d;
-import edu.wpi.first.networktables.NetworkTable;
-import edu.wpi.first.networktables.NetworkTableEntry;
-import edu.wpi.first.networktables.NetworkTableInstance;
-import frc.robot.Constants;
-import frc.robot.Constants.Vision;
-import frc.robot.LimelightHelpers;
-
-/** Add your docs here. */
-public class AprilTag {
-
-  private class TargetVector {
-
-  }
-
-    public static final NetworkTable TABLE = NetworkTableInstance.getDefault().getTable("limelight");
-
-    public static NetworkTableEntry entry(String key) {
-        return TABLE.getEntry(key);
-    }
-
-    public boolean hasTarget() {
-        return targetPos() != null;
-    }
-
-    public static void aprilTagPeriodic() {
-      var targetPosition = targetPos();
-      // Transform3d relativeRobotPosition;
-      if (targetPosition == null) {
-        System.out.println("No Limelight target.");
-      } else {
-        // System.out.printf("Target position: {x: %.3f, y: %.3f, z: %.3f}\n", t.getX(), t.getY(), t.getZ());
-  
-        if (AprilTag.getDirectDistance() != -1) {
-          System.out.println("Current spatial distance to primary target: " + AprilTag.getDirectDistance());
-        }
-      }
-    }
-
-  // private static double tx, ty, height;
-  private static double angle;
-  private static double height;
-
-  // public AprilTag(double tx, double ty, double height) {
-  //   this.tx = tx;
-  //   this.ty = ty;
-  //   this.height = height;
-  // }
-
-//   public static double[] targetToXYZVector() {
-//     double[] targetVector = new double[3];
-//     var currentTargetID = Limelight.entry("tid").getInteger(-1);
-//     if (currentTargetID != -1) {
-//      try {
-//       // height = Constants.APRILTAGS.get(currentTargetID).getZ();
-//       height = 0.3429;
-//      } catch (Exception systemException) {}
-//     } else {
-//       System.out.print("Unfortunately, no limelight is detected.");
-//       return null;
-//     }
-
-//     double z = height - Vision.kLimelightHeightMeters;
-//     double y = z / Math.tan(Math.toRadians(LimelightHelpers.getTY("") + Vision.kLimelightAngleDegrees)); // also need Math.abs()?
-//     // double y = z / Math.tan(Math.toRadians(Limelight.entry("ty").getDouble(-1) + Vision.kLimelightAngleDegrees)); // also need Math.abs()?
-
-//     double x = y * Math.tan(Math.toRadians(LimelightHelpers.getTX("")));
-//     // double x = y * Math.tan(Math.toRadians(Limelight.entry("tx").getDouble(-1)));
-
-//     targetVector[0] = x;
-//     targetVector[1] = y;
-//     targetVector[2] = z;
-
-//     return targetVector;
-//   }
-
-  public static Transform3d targetPos() {
-        var targetPose = entry("targetpose_robotspace").getDoubleArray(new double[0]);
-        if (targetPose.length != 6 || targetPose[2] < 1E-6) return null;
-        
-        return new Transform3d(
-            new Translation3d(targetPose[0], targetPose[1], targetPose[2]),
-            new Rotation3d(targetPose[3], targetPose[4], targetPose[5])
-        );
-    };
-
-
-  public static double getDirectDistance() {
-    Transform3d targetVector = targetPos();
-
-    if (targetPos() != null) {
-      return Math.sqrt(Math.pow(targetVector.getX(), 2) + Math.pow(targetVector.getY(), 2) + Math.pow(targetVector.getZ(), 2));
-    } else {
-      return -1;
-    }
-  }
-
-    // The method declared below exists as a way in which one is able to determine the robot's current location and orientation within the playing field shown within the FIRST Robotics 2024 competition, Crescendo.
-
-  public static Transform3d determinePosition() {
-    var relativePositionalData = targetPos();
-    var aprilTagIdentifier = entry("tid").getInteger(-1);
-    if (aprilTagIdentifier == -1 || aprilTagIdentifier > 16) {
-        return null;
-    }
-    Transform3d currentAprilTagLocation = Constants.APRILTAGS.get(aprilTagIdentifier);
-
-    double[] currentRelativeLocation = {
-        currentAprilTagLocation.getX() + relativePositionalData.getX(), 
-        currentAprilTagLocation.getY() + relativePositionalData.getY(), 
-        currentAprilTagLocation.getZ() + relativePositionalData.getZ()
-    };
-
-    double[] currentRelativeRotation = {
-        currentAprilTagLocation.getRotation().getX() + relativePositionalData.getRotation().getX(),
-        currentAprilTagLocation.getRotation().getY() + relativePositionalData.getRotation().getY(),
-        currentAprilTagLocation.getRotation().getZ() + relativePositionalData.getRotation().getZ(),
-    };
-
-    Transform3d currentRelativePosition = new Transform3d (
-        new Translation3d(currentRelativeLocation[0], currentRelativeLocation[1], currentRelativeLocation[2]),
-        new Rotation3d(currentRelativeRotation[0], currentRelativeRotation[1], currentRelativeRotation[2])
-    );
-
-    System.out.printf("Current relative robot position: {X: %0.3f, Y: %0.3f, Z: %0.3f}\n", currentRelativePosition.getX(), currentRelativePosition.getY(), currentRelativePosition.getZ());
-    System.out.printf("Current relative robot orientation: {X: %0.3f, Y: %0.3f, Z: %0.3f}\n", currentRelativePosition.getRotation().getX(), currentRelativePosition.getRotation().getY(), currentRelativePosition.getRotation().getZ());
-
-    return currentRelativePosition;
-}
-
-public Rotation2d determineTargetRotationalOffset(Translation3d targetPosition) {
-        // Transform3d currentRelativePosition = determineRelativePosition();
-        Transform3d currentRelativePosition = new Transform3d (
-            new Translation3d(1, 1, 1),
-            new Rotation3d(1, 1, 1)
-        );
-
-        if (currentRelativePosition == null) {
-            System.out.printf("An apriltag is currently not within view, and therefore, we are unable to calculate the positional offset of the target located at {X: %.3f, Y: %.3f, Z: %.3f}\n.", targetPosition.getX(), targetPosition.getY(), targetPosition.getZ());     
-            return null;
-        }
-
-        if ((currentRelativePosition.getX() == targetPosition.getX()) && (currentRelativePosition.getY() == targetPosition.getY())) {
-            System.out.println("As the given position is the one in which the robot is currently located, we are unfortunately, unable to calculate the rotational offset to said location.");
-            return null;
-        }
-
-        double targetXAxisOffset = Math.toDegrees(Math.asin(Math.abs(currentRelativePosition.getX() - targetPosition.getX()) / Math.sqrt(Math.pow(currentRelativePosition.getX() - targetPosition.getX(), 2) + Math.pow(currentRelativePosition.getY() - targetPosition.getY(), 2))));
-        double targetZAxisOffset;
-
-        if (currentRelativePosition.getZ() == targetPosition.getZ()) {
-            targetZAxisOffset = 0;
-        } else {
-            targetZAxisOffset = Math.toDegrees(Math.asin(Math.abs(currentRelativePosition.getZ() - targetPosition.getZ()) / (targetPosition.getDistance(currentRelativePosition.getTranslation()))));
-        }
-
-        System.out.println("Target X Offset: " + targetXAxisOffset + ", Target Z Offset: " + targetZAxisOffset);
-        
-        return new Rotation2d(targetXAxisOffset, targetZAxisOffset);
-    }
-
-
-
-  // public double[] targetToXYZVector() {
-  // double[] targetVector = new double[3];
-
-  // double d = getDirectDistance();
-
-  // double x = d * Math.sin(LimelightHelpers.getTX(""));
-  // targetVector[0] = x;
-
-  // return targetVector;
-  // }
-
-  // public double getDirectDistance() {
-  // return (/*placeholder*/3);
-  // }
-
-}
