--- SY2024-CORE-A-master/src/main/java/frc/robot/Main.java
@@ -1,19 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot;
-
-import edu.wpi.first.wpilibj.RobotBase;
-
-/**
- * The main class for the robot.
- * This simply calls a WPILib function to initialize the robot.
- */
-public final class Main {
-  private Main() {}
-
-  public static void main(String... args) {
-    RobotBase.startRobot(Robot::new);
-  }
-}
