--- SY2024-CORE-A-master/src/main/java/frc/robot/Robot.java
@@ -1,88 +1,0 @@
-// Copyright (c) FIRST and other WPILib contributors.
-// Open Source Software; you can modify and/or share it under the terms of
-// the WPILib BSD license file in the root directory of this project.
-
-package frc.robot;
-
-import edu.wpi.first.wpilibj.TimedRobot;
-import edu.wpi.first.wpilibj2.command.Command;
-import edu.wpi.first.wpilibj2.command.CommandScheduler;
-import frc.lib.config.CTREConfigs;
-import frc.lib.util.AprilTag;
-import frc.robot.subsystems.Intake;
-
-/**
- * The main robot class. This handles the robot container (which contains the
- * subsystems) and handles triggers during different modes.
- */
-public class Robot extends TimedRobot {
-  public static CTREConfigs ctreConfigs;
-  private Command m_autonomousCommand;
-
-  public long initalizationTime;
-  private RobotContainer m_robotContainer;
-  // private AprilTag aprilTagInstance;
-
-  @Override
-  public void robotInit() {
-    ctreConfigs = new CTREConfigs();
-    m_robotContainer = new RobotContainer();
-    // aprilTagInstance = new AprilTag();
-    initalizationTime = System.currentTimeMillis();
-  }
-
-  @Override
-  public void robotPeriodic() {
-    CommandScheduler.getInstance().run();
-    AprilTag.aprilTagPeriodic();
-    Intake.intakePeriodic();
-  }
-
-  @Override
-  public void disabledInit() {}
-
-  @Override
-  public void disabledPeriodic() {}
-
-  @Override
-  public void disabledExit() {}
-
-  @Override
-  public void autonomousInit() {
-    m_autonomousCommand = m_robotContainer.getAutonomousCommand();
-
-    if (m_autonomousCommand != null) {
-      m_autonomousCommand.schedule();
-    }
-  }
-
-  @Override
-  public void autonomousPeriodic() {}
-
-  @Override
-  public void autonomousExit() {}
-
-  @Override
-  public void teleopInit() {
-    if (m_autonomousCommand != null) {
-      m_autonomousCommand.cancel();
-    }
-  }
-
-  @Override
-  public void teleopPeriodic() {}
-
-  @Override
-  public void teleopExit() {}
-
-  @Override
-  public void testInit() {
-    CommandScheduler.getInstance().cancelAll();
-  }
-
-  @Override
-  public void testPeriodic() {}
-
-  @Override
-  public void testExit() {}
-}
