--- SY2024-CORE-A-master/src/main/java/frc/robot/subsystems/Swerve.java
@@ -1,142 +1,0 @@
-package frc.robot.subsystems;
-
-
-import com.ctre.phoenix6.configs.Pigeon2Configuration;
-import com.ctre.phoenix6.configs.TalonFXConfiguration;
-import com.ctre.phoenix6.hardware.Pigeon2;
-import com.pathplanner.lib.auto.AutoBuilder;
-import com.pathplanner.lib.util.HolonomicPathFollowerConfig;
-import com.pathplanner.lib.util.PIDConstants;
-import com.pathplanner.lib.util.ReplanningConfig;
-
-import edu.wpi.first.math.geometry.Pose2d;
-import edu.wpi.first.math.geometry.Rotation2d;
-import edu.wpi.first.math.geometry.Translation2d;
-import edu.wpi.first.math.kinematics.ChassisSpeeds;
-import edu.wpi.first.math.kinematics.SwerveDriveKinematics;
-import edu.wpi.first.math.kinematics.SwerveDriveOdometry;
-import edu.wpi.first.math.kinematics.SwerveModulePosition;
-import edu.wpi.first.math.kinematics.SwerveModuleState;
-import edu.wpi.first.wpilibj.DriverStation;
-import edu.wpi.first.wpilibj.smartdashboard.Field2d;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
-import edu.wpi.first.wpilibj2.command.SubsystemBase;
-import frc.robot.Constants;
-
-public class Swerve extends SubsystemBase {
-  private final Pigeon2 gyro;
-
-  private SwerveDriveOdometry swerveOdometry;
-  private SwerveModule[] mSwerveMods;
-
-  private Field2d field;
-
-  public Swerve() {
-    gyro = new Pigeon2(Constants.Swerve.pigeonID);
-    gyro.getConfigurator().apply(new Pigeon2Configuration());
-    zeroGyro();
-
-    mSwerveMods =
-        new SwerveModule[] {
-          new SwerveModule(0, Constants.Swerve.Mod0.constants),
-          new SwerveModule(1, Constants.Swerve.Mod1.constants),
-          new SwerveModule(2, Constants.Swerve.Mod2.constants),
-          new SwerveModule(3, Constants.Swerve.Mod3.constants)
-        };
-
-    swerveOdometry = new SwerveDriveOdometry(Constants.Swerve.swerveKinematics, getYaw(), getModulePositions());
-
-    var pathConfig = new HolonomicPathFollowerConfig(
-        new PIDConstants(Constants.PathPlanner.driveKP, Constants.PathPlanner.driveKI, Constants.PathPlanner.driveKD),
-        new PIDConstants(Constants.PathPlanner.turnKP, Constants.PathPlanner.turnKI, Constants.PathPlanner.turnKD),
-        Constants.Swerve.maxSpeed,
-        Constants.Swerve.wheelBase,
-        new ReplanningConfig());
-
-    AutoBuilder.configureHolonomic(
-        this::getPose,
-        this::resetOdometry,
-        () -> Constants.Swerve.swerveKinematics.toChassisSpeeds(getStates()),
-        speeds -> setModuleStates(Constants.Swerve.swerveKinematics.toSwerveModuleStates(speeds)),
-        pathConfig,
-        () -> DriverStation.getAlliance().filter(a -> a == DriverStation.Alliance.Red).isPresent(),
-        this);
-
-    field = new Field2d();
-    SmartDashboard.putData("Field", field);
-  }
-
-  public void drive(
-      Translation2d translation, double rotation, boolean fieldRelative, boolean isOpenLoop) {
-    SwerveModuleState[] swerveModuleStates =
-        Constants.Swerve.swerveKinematics.toSwerveModuleStates(
-            fieldRelative
-                ? ChassisSpeeds.fromFieldRelativeSpeeds(
-                    translation.getX(), translation.getY(), rotation, getYaw())
-                : new ChassisSpeeds(translation.getX(), translation.getY(), rotation));
-    SwerveDriveKinematics.desaturateWheelSpeeds(swerveModuleStates, Constants.Swerve.maxSpeed);
-
-    for (SwerveModule mod : mSwerveMods) {
-      mod.setDesiredState(swerveModuleStates[mod.moduleNumber], isOpenLoop);
-    }
-  }
-
-  /* Used by SwerveControllerCommand in Auto */
-  public void setModuleStates(SwerveModuleState[] desiredStates) {
-    SwerveDriveKinematics.desaturateWheelSpeeds(desiredStates, Constants.Swerve.maxSpeed);
-
-    for (SwerveModule mod : mSwerveMods) {
-      mod.setDesiredState(desiredStates[mod.moduleNumber], false);
-    }
-  }
-
-  public Pose2d getPose() {
-    return swerveOdometry.getPoseMeters();
-  }
-
-  public SwerveModulePosition[] getModulePositions() {
-    SwerveModulePosition[] positions = new SwerveModulePosition[4];
-
-    for(int i = 0; i < mSwerveMods.length; i++) {
-      positions[i] = mSwerveMods[i].getPosition();
-    }
-    return positions;
-  }
-
-  public void resetOdometry(Pose2d pose) {
-    swerveOdometry.resetPosition(getYaw(), getModulePositions(), pose);
-  }
-
-  public SwerveModuleState[] getStates() {
-    SwerveModuleState[] states = new SwerveModuleState[4];
-    for (SwerveModule mod : mSwerveMods) {
-      states[mod.moduleNumber] = mod.getState();
-    }
-    return states;
-  }
-
-  public void zeroGyro() {
-    gyro.setYaw(0);
-  }
-
-  public Rotation2d getYaw() {
-    return (Constants.Swerve.invertGyro)
-        ? Rotation2d.fromDegrees(360 - gyro.getYaw().getValue())
-        : Rotation2d.fromDegrees(gyro.getYaw().getValue());
-  }
-
-  @Override
-  public void periodic() {
-    swerveOdometry.update(getYaw(), getModulePositions());
-    field.setRobotPose(getPose());
-
-    for (SwerveModule mod : mSwerveMods) {
-      SmartDashboard.putNumber(
-          "Mod " + mod.moduleNumber + " Cancoder", mod.getCanCoder().getDegrees());
-      SmartDashboard.putNumber(
-          "Mod " + mod.moduleNumber + " Integrated", mod.getState().angle.getDegrees());
-      SmartDashboard.putNumber(
-          "Mod " + mod.moduleNumber + " Velocity", mod.getState().speedMetersPerSecond);
-    }
-  }
-}
