--- SY2024-CORE-A-master/vendordeps/PathplannerLib.json
@@ -1,38 +1,0 @@
-{
-    "fileName": "PathplannerLib.json",
-    "name": "PathplannerLib",
-    "version": "2024.1.3",
-    "uuid": "1b42324f-17c6-4875-8e77-1c312bc8c786",
-    "frcYear": "2024",
-    "mavenUrls": [
-        "https://3015rangerrobotics.github.io/pathplannerlib/repo"
-    ],
-    "jsonUrl": "https://3015rangerrobotics.github.io/pathplannerlib/PathplannerLib.json",
-    "javaDependencies": [
-        {
-            "groupId": "com.pathplanner.lib",
-            "artifactId": "PathplannerLib-java",
-            "version": "2024.1.3"
-        }
-    ],
-    "jniDependencies": [],
-    "cppDependencies": [
-        {
-            "groupId": "com.pathplanner.lib",
-            "artifactId": "PathplannerLib-cpp",
-            "version": "2024.1.3",
-            "libName": "PathplannerLib",
-            "headerClassifier": "headers",
-            "sharedLibrary": false,
-            "skipInvalidPlatforms": true,
-            "binaryPlatforms": [
-                "windowsx86-64",
-                "linuxx86-64",
-                "osxuniversal",
-                "linuxathena",
-                "linuxarm32",
-                "linuxarm64"
-            ]
-        }
-    ]
-}
