+++ src/main/java/frc/robot/Constants.java
@@ -38,6 +38,15 @@ public final class Constants {
     public static final double kD = 0.0;
   }
 
+  public static final class Intake {
+    public static final double intakeMotorVelocity = 0;
+    public static final double intakeFlywheelRadius = 0;
+    public static final double gravityAcceleration = 0;
+    public static final double kP = 0;
+    public static final double kI = 0;
+    public static final double kD = 0;
+  }
+
   public static class MoveTowardTag {
 
     // Measured in portion of maximum speed
